pyproject.toml
[project]
name = "jira-summary"
version = "0.1.0"
description = ""
authors = [
    {name = "Karl Boehnker",email = "karlwb@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "tomlkit (>=0.13.3,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
